+++ python/lsst/ts/ess/csc/data_client/young_32400_weather_station_data_client.py
@@ -259,7 +259,7 @@ class Young32400WeatherStationDataClient(common.BaseDataClient):
         return yaml.safe_load(
             """
 $schema: http://json-schema.org/draft-07/schema#
-description: Schema for RPiDataClient
+description: Schema for Young32400WeatherStationDataClient.
 type: object
 properties:
   host: